Add files via upload
Internship play a really important role in a student's life as they introduce them to the actual working of an  industry. Getting a good internship is like an achievement for every student. 
So to aid the students in getting the better internship, I have created an automation project which helps in building your resume on one of the greatest internship site: "Internshala". 
This project also helps in applying for some of the internship based on your resume.